refactor(models): move volatility contracts
